Set the timezone to UTC
The timezone for the timestamps should be universal. Otherwise, there can be mismatches due to the different offsets, causing some hosts to think others are dead, login sessions to expire at different times according to different hosts, etc.

As for why UTC isn't used by default and people have to find this out the hard way, who knows.

Also, it would be nice to set the timestamp column data type to timestamp with timezone so that it can not only be consistent but also be interpreted properly, but H2 decides to fail for some reason (tried any combination of Instant/OffsetDateTime + timestamp/timestamptz).

No need to deploy this to the server - the UTC timezone is already used there.