Redesign the host list
Now the host list is a separate page with an actual Material list (instead of an accordion) and clicking on the name of a host leads to another page, which is the host details (Terraria worlds and instances).

The host list items look much cleaner now, without the clutter of the details and the uncomfortable accordion getting in the way.