Remove the JUnit report plugin from the pipeline
Seriously... this plugin does not understand the concept of not rerunning the tests if the source code hasn't changed and doesn't even let me tell it not to fail with "Test reports were found but none of them are new". I don't need it anyway since I already have the JaCoCo code coverage chart.