[Close #28] Execute instance actions as soon as possible
Use TerrariaInstanceSubscriptionService to keep an eye on all instances. This is much better than having a manually managed BlockingQueue in TerrariaInstanceActionService itself. It prevents other services from becoming coupled with it, and it makes the instance tracking more agnostic in the sense that later the PostgreSQL notification mechanism can be implemented more easily.