Use `null` for the nullable types coming from the backend
For now `null` and `undefined` are interchangeable as far as the frontend is concerned, but it's nice for it to know exactly what kind of properties come from the backend regardless.

Also, the `@Nullable` annotation of Checker Framework can be applied at the type level, where "optional property" can't always apply.