[Close #48] Replace Checker Framework with SpotBugs
However, keep using the Checker Framework annotations because they're the only ones that are actually good for my purposes (specifically, annotating types).

I will not use any SpotBugs or JSR305 annotations. The nullability checks of SpotBugs result in a lot of false negatives anyway (not failing when it should fail). Regardless, SpotBugs is now a Gradle task in the backend project - `:backend:spotbugsMain`.