[Close #29] Watch for PostgreSQL notifications
In order to do so, use a new PostgreSQL driver - `com.impossibl.postgres.jdbc.PGDriver`. The old one - `org.postgresql.Driver` - was fine too, but it's much harder to fetch the notifications for it properly because it isn't handler-based, and instead you're forced to poll the notifications. That doesn't result in database spam, but it does use the CPU unnecessarily.